Modify Area and ServiceBudget in order to link them properly and be able to use them in dropdowns from the web-app.

+++ src/main/java/com/insside/budget/area/AreaRepository.java
@@ -2,6 +2,6 @@ package com.insside.budget.area;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AreaRepository extends JpaRepository<Area, Integer> {
+public interface AreaRepository extends JpaRepository<Area, Long> {
 
 }
